refactor: ReasonPhraseUtil로 추출
